+++ api/dashboard/src/app/login/login.component.html
@@ -3,11 +3,11 @@
     <img class="mb-4" src="assets/img/cloud.png" alt="" height="57" />
     <h1 class="h3 mb-3 fw-normal">Please sign in</h1>
     <input
-    #email
-      type="email"
-      id="inputEmail"
+    #username
+      type="username"
+      id="inputUsername"
       class="form-control"
-      placeholder="Email address"
+      placeholder="Username"
       required
       autofocus
     />
@@ -19,7 +19,7 @@
       placeholder="Password"
       required
     />
-    <button class="w-100 btn btn-lg btn-primary" (click)="login(email.value, password.value)">Sign in</button>
+    <button class="w-100 btn btn-lg btn-primary" (click)="login(username.value, password.value)">Sign in</button>
     <p class="mt-5 mb-3 text-muted">&copy; 2017–2021</p>
   </main>
 </div>
