+++ api/v1/commands.js
@@ -1,0 +1,101 @@
+const express = require("express");
+const router = express.Router();
+const expressJwt = require("express-jwt");
+const spawn = require("child_process").spawn;
+
+var processingCommand = false;
+
+function isCommandRunning(req, res, next) {
+  if (processingCommand) {
+    res.status(409).send("command processing...");
+  } else {
+    next();
+  }
+}
+
+//scale up instance
+router.post(
+  "/up",
+  expressJwt({ secret: process.env.JWT_SECRET, algorithms: ["HS256"] }),
+  isCommandRunning,
+  (req, res) => {
+    runCommand("ls", req.app.get('socketio'));
+    res.sendStatus(200);
+  }
+);
+
+//scale down instance
+router.post(
+  "/down",
+  expressJwt({ secret: process.env.JWT_SECRET, algorithms: ["HS256"] }),
+  isCommandRunning,
+  (req, res) => {
+    res.sendStatus(200);
+  }
+);
+
+//setup - would require post data for all of the options and to modify the script to accept flags
+router.post(
+  "/setup",
+  expressJwt({ secret: process.env.JWT_SECRET, algorithms: ["HS256"] }),
+  isCommandRunning,
+  (req, res) => {
+    res.sendStatus(200);
+  }
+);
+
+//change region
+router.post(
+  "/region",
+  expressJwt({ secret: process.env.JWT_SECRET, algorithms: ["HS256"] }),
+  isCommandRunning,
+  (req, res) => {
+    res.sendStatus(200);
+  }
+);
+
+//get status
+router.post(
+  "/status",
+  expressJwt({ secret: process.env.JWT_SECRET, algorithms: ["HS256"] }),
+  isCommandRunning,
+  (req, res) => {
+    res.sendStatus(200);
+  }
+);
+
+//nuke and run
+router.post(
+  "/destroy",
+  expressJwt({ secret: process.env.JWT_SECRET, algorithms: ["HS256"] }),
+  isCommandRunning,
+  (req, res) => {
+    res.sendStatus(200);
+  }
+);
+
+router.use(function (err, req, res, next) {
+  if (err.name === "UnauthorizedError") {
+    res.status(401).send("invalid token...");
+  }
+});
+
+function runCommand(command, io) {
+  processingCommand = true;
+  var myProcess = spawn(command);
+  myProcess.stdout.setEncoding("utf-8");
+  myProcess.stdout.on("data", function (data) {
+    io.sockets.emit("process_data", data);
+  });
+
+  myProcess.stderr.setEncoding("utf-8");
+  myProcess.stderr.on("error", function (data) {
+    io.sockets.emit("process_data", data);
+  });
+
+  myProcess.on("exit", () => {
+    processingCommand = false;
+  });
+}
+
+module.exports = router;
